Add files via upload
V ehicle insurance fraud refers to the intentional act of cheating an auto insurance provider to
obtain benefits or compensation to which the claimant is not legitimately entitled. This activity is
a growing problem worldwide and can be difficult to detect. Insurance companies lose millions
of dollars each year due to fraudulent activities related to auto insurance fraud. Insurance fraud
comes in two main forms: hard fraud and soft fraud. Hard fraud is intentional damage caused in
order to pass it off as an insurance loss. Examples include staged accidents and false reports of a
stolen car. On the other hand, Soft car insurance fraud occurs when a legitimate claim is
submitted, but damages or injuries are overstated in an attempt to secure a higher payout. This
may involve practices such as over-valuing, exaggerating claims and misrepresentations.
Insurance companies employ advanced data analytics and rigorous investigations to proactively
detect and prevent fraudulent activities. By leveraging advanced technologies, we can identify
patterns and anomalies, minimize financial losses, protect the integrity of the insurance system
and ensure fair treatment for all policyholders